backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai>=0.3.0>=0.3.0
sentence-transformers==2.2.2
numpy==1.24.3
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiosmtplib==3.0.1
email-validator==2.1.0
